odyssey-intro-strawberry/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
define==1.74
fastapi==0.112.0
graphql-core==3.2.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
mypy==1.11.1
mypy-extensions==1.0.0
packaging==24.1
pluggy==1.5.0
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
strawberry-graphql==0.237.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
